Configures devise validatable
Fixes #70

Configures User model to work with Devise validatable. This
enables Devise to compare the password and password_confirmation
fields in the registration, password change and password reset
controller.

**NOTE**
This also configures a minimum password length of 6 characters,
and a maximum password length of 128 characters.
Existing users with shorter passwords continue to be valid, and
can still login with their existing passwords.